+++ Internal/PrefabSafeSet/Runtime/PrefabSafeSet.cs
@@ -38,7 +38,7 @@ namespace Anatawa12.AvatarOptimizer.PrefabSafeSet
             foreach (var assembly in AppDomain.CurrentDomain.GetAssemblies())
             {
                 OnBeforeSerializeImplType =
-                    assembly.GetType("Anatawa12.AvatarOptimizer.PrefabSafeSet.OnBeforeSerializeImpl`1");
+                    assembly.GetType("Anatawa12.AvatarOptimizer.PrefabSafeSet.PrefabSafeSetRuntimeEditorImpl`1");
                 if (OnBeforeSerializeImplType != null) return;
             }
             if (OnBeforeSerializeImplType == null)
@@ -48,7 +48,14 @@ namespace Anatawa12.AvatarOptimizer.PrefabSafeSet
         public static MethodInfo GetOnBeforeSerializeCallbackMethod(Type tType, Type setType)
         {
             var implType = OnBeforeSerializeImplType.MakeGenericType(tType);
-            return implType.GetMethod("Impl", BindingFlags.Public | BindingFlags.Static, null, new[] { setType }, null)!;
+            return implType.GetMethod("OnBeforeSerialize", BindingFlags.Public | BindingFlags.Static, null, new[] { setType }, null)!;
+        }
+
+        public static MethodInfo GetOnValidateCallbackMethod(Type tType, Type tComponentType)
+        {
+            var implType = OnBeforeSerializeImplType.MakeGenericType(tType);
+            return implType.GetMethod("OnValidate", BindingFlags.Public | BindingFlags.Static)!
+                .MakeGenericMethod(tComponentType);
         }
 #endif
     }
@@ -70,18 +77,21 @@ namespace Anatawa12.AvatarOptimizer.PrefabSafeSet
     /// </summary>
     /// <typeparam name="T">Element Type</typeparam>
     [NotKeyable, Serializable]
-    public class PrefabSafeSet<T> : PrefabSafeSetApi<T>, ISerializationCallbackReceiver
+    public class PrefabSafeSet<T> : PrefabSafeSetApi<T>
     {
         [SerializeField] internal T[] mainSet = Array.Empty<T>();
         [SerializeField] internal PrefabLayer<T>[] prefabLayers = Array.Empty<PrefabLayer<T>>();
+        // If the PrefabSafeSet is on scene and prefab instance, this will be used
+        // This is added AAO 1.8.0 to support replacing base prefab on the scene, since Unity 2022
+        [SerializeField] internal bool usingOnSceneLayer;
+        [SerializeField] internal PrefabLayer<T> onSceneLayer = new();
 
 #if UNITY_EDITOR
         [SerializeField, HideInInspector] internal T? fakeSlot;
-        internal readonly Object OuterObject;
-        internal T[]? CheckedCurrentLayerRemoves;
-        internal T[]? CheckedCurrentLayerAdditions;
-        private static MethodInfo _onBeforeSerializeCallback = PrefabSafeSetRuntimeUtil
-            .GetOnBeforeSerializeCallbackMethod(typeof(T), typeof(PrefabSafeSet<T>));
+        internal Object OuterObject;
+        internal Object? CorrespondingObject;
+        internal int? NestCount;
+        internal bool IsNew;
 #endif
 
         public PrefabSafeSet(Object outerObject)
@@ -91,6 +101,8 @@ namespace Anatawa12.AvatarOptimizer.PrefabSafeSet
             // so use actual null check instead of destroy check
             // ReSharper disable once Unity.NoNullCoalescing
             OuterObject = outerObject ?? throw new ArgumentNullException(nameof(outerObject));
+            IsNew = true;
+            UnityEditor.EditorApplication.delayCall += () => IsNew = false;
 #endif
         }
 
@@ -235,18 +247,23 @@ namespace Anatawa12.AvatarOptimizer.PrefabSafeSet
         public override void RemoveIf(Func<T, bool> predicate) => throw new Exception("Not supported in Player build");
         public override void Clear() => throw new Exception("Not supported in Player build");
 #endif
+    }
 
-        void ISerializationCallbackReceiver.OnBeforeSerialize()
+    public static class PrefabSafeSet {
+        public static void OnValidate<T, TComponent>(TComponent component, Func<TComponent, PrefabSafeSet<T>> getPrefabSafeSet) where TComponent : Component
         {
 #if UNITY_EDITOR
-            _onBeforeSerializeCallback.Invoke(null, new object[] {this});
+            ValidateMethodHolder<T, TComponent>.OnValidateCallbackMethodGeneric.Invoke(null,
+                new object[] { component, getPrefabSafeSet });
 #endif
         }
-
-        void ISerializationCallbackReceiver.OnAfterDeserialize()
+#if UNITY_EDITOR
+        private static class ValidateMethodHolder<T, TComponent>
         {
-            // there's nothing to do after deserialization.
+            public static MethodInfo OnValidateCallbackMethodGeneric =
+                PrefabSafeSetRuntimeUtil.GetOnValidateCallbackMethod(typeof(T), typeof(TComponent));
         }
+#endif
     }
 
     [Serializable]